Moved repository interfaces to repository folder
