+++ KanbanCord.Bot/Commands/ArchiveCommand.cs
@@ -4,8 +4,8 @@ using DSharpPlus.Commands.Processors.SlashCommands;
 using DSharpPlus.Entities;
 using KanbanCord.Bot.Extensions;
 using KanbanCord.Bot.Helpers;
-using KanbanCord.Core.Interfaces;
 using KanbanCord.Core.Models;
+using KanbanCord.Core.Repositories;
 
 namespace KanbanCord.Bot.Commands;
