+++ KanbanCord.Bot/Commands/BoardCommand.cs
@@ -4,8 +4,8 @@ using DSharpPlus.Commands.Processors.SlashCommands;
 using DSharpPlus.Commands.Processors.SlashCommands.ArgumentModifiers;
 using KanbanCord.Bot.Helpers;
 using KanbanCord.Bot.Providers;
-using KanbanCord.Core.Interfaces;
 using KanbanCord.Core.Models;
+using KanbanCord.Core.Repositories;
 
 namespace KanbanCord.Bot.Commands;
