+++ KanbanCord.Bot/Commands/ClearCommand.cs
@@ -5,8 +5,8 @@ using DSharpPlus.Commands.Processors.SlashCommands;
 using DSharpPlus.Entities;
 using DSharpPlus.Interactivity.Extensions;
 using KanbanCord.Bot.Extensions;
-using KanbanCord.Core.Interfaces;
 using KanbanCord.Core.Models;
+using KanbanCord.Core.Repositories;
 
 namespace KanbanCord.Bot.Commands;
