+++ KanbanCord.Bot/Commands/ResetCommand.cs
@@ -5,7 +5,7 @@ using DSharpPlus.Commands.Processors.SlashCommands;
 using DSharpPlus.Entities;
 using DSharpPlus.Interactivity.Extensions;
 using KanbanCord.Bot.Extensions;
-using KanbanCord.Core.Interfaces;
+using KanbanCord.Core.Repositories;
 
 namespace KanbanCord.Bot.Commands;
