+++ KanbanCord.Bot/EventHandlers/GuildDeletedEventHandler.cs
@@ -1,6 +1,6 @@
 using DSharpPlus;
 using DSharpPlus.EventArgs;
-using KanbanCord.Core.Interfaces;
+using KanbanCord.Core.Repositories;
 using Microsoft.Extensions.Logging;
 
 namespace KanbanCord.Bot.EventHandlers;
