+++ KanbanCord.Bot/Extensions/ServiceCollectionExtensions.cs
@@ -5,7 +5,6 @@ using DSharpPlus.Extensions;
 using DSharpPlus.Interactivity.Extensions;
 using KanbanCord.Bot.BackgroundServices;
 using KanbanCord.Bot.EventHandlers;
-using KanbanCord.Core.Interfaces;
 using KanbanCord.Core.Options;
 using KanbanCord.Core.Repositories;
 using Microsoft.Extensions.Options;
