+++ KanbanCord.Bot/Providers/AllTaskItemsAutoCompleteProvider.cs
@@ -2,7 +2,7 @@ using DSharpPlus.Commands.Processors.SlashCommands;
 using DSharpPlus.Commands.Processors.SlashCommands.ArgumentModifiers;
 using DSharpPlus.Entities;
 using KanbanCord.Bot.Helpers;
-using KanbanCord.Core.Interfaces;
+using KanbanCord.Core.Repositories;
 
 namespace KanbanCord.Bot.Providers;
